requirements.txt
django==4.0.4
djangorestframework
pyyaml
requests
django-cors-headers
pillow